type of return is changed from integer to long

+++ src/main/java/uz/yeoju/yeoju_app/payload/resDto/rektor/dashboard/RektorDashboard.java
@@ -26,7 +26,7 @@ public interface RektorDashboard {
     List<StaffCountComeAndAll> getStaffs();
 
     @Value("#{@studentRepository.getStudentComeCount()}")
-    List<Integer> getStudents();
+    List<Long> getStudents();
 
 //    @Value("#{@staffRepository.getStaffComeCountStatistics()}")
 //    List<TeacherCountComeAndAll> getStaffsStatistics();
